test(scm): added TestMockCreateGroup and revised CreateGroup
CreateGroup now uses a new getRepository helper function to check
if the group's repository already exists on the SCM, and returns
an appropriate error message if it exists.

This needs a few minor tweaks to TestMockGetRepositories and
TestMockCreateCourse since the state of the mock implementation
now adds repos and collaborators.