refactor(scm): RepositoryIsEmpty now first checks for org/repo
This is a minor tweak to check if org/repo exists before checking
whether or not the repo contains any content. The actual content
check is the same as before, but it isn't clear if the rather
complex check is necessary.

This also adds some additional logging to allow us to better
understand real-world scenarios.