fix(scm): removed slug.Make(opt.Path) in createRepository
Actually, the slug.Make() was only used to check for existing
repository with the same name. This would have failed, if first
a non-slugged repo name were used to create the repo followed by
another attempt to create the same repo, but this time checking
with a slugged version of the repo name; which wouldn't find the
non-slugged version. This in turn would have attempted to create
another non-slugged repo, and thus would have failed to create
the repo and not return the scm.Repository struct that the current
implementation would have done for a slug-compatible name.

We should instead consider to always pass a slug-compatible name
to the SCM methods, and moreover, we should only store the "slugged"
name in the database as well, if we do that.