fix(scm): TestPromoteDemoteRejectTeacher now works with new mock
We need to enroll students properly via the client instead of
cheating by only adding enrollments in the db; otherwise, the mock
SCM will not know about the enrolled students and thus cannot
promote them to teachers.

This also adds the WithMockOrgs() helper to fill the orgs slice
corresponding to the qtest.MockCourses.