chore(scm): reorganized scm method order to be consistent
This organizes the methods in the SCM interface to keep relevant
methods together; and this also reorganizes the tests ordering and
implementation ordering accordingly.

PS: This revealed a bug in the mock that requires that the input
default groups much be deep cloned (will be fixed in a follow up
commit). There may be more bugs in the mock for similar reasons,
but will fix them as we discover them. That said, we could also
try to deep clone in all With* mock options to be safe.