refactor(scm): WithMockOrgs(members ...string)
This adds support for adding members to the mock organizations via
the variadic members list; the first member is assigned as owner,
and the rest as regular members. Currently, only owner is used in
actual tests.

Along with this change, web.MockClient's signature has also been
updated to accept a scm.MockOption instead of only using the
hard-coded one (WithMockOrgs()).

Finally, the MockManager now initializes the scms map for each
course organization specified in the mock's orgs list so that
each course/organization lookup gets a valid scm instance.