refactor(scm): SCMError now uses a E() and M() funcs
This commit contains too many changes to describe in detail;
mainly it changes error messages, but makes a number of other
improvements as well, including renaming grantAccess to addUser()
to make it clear that this also adds the user to the repo.

The changes in github_issues.go does not have tests for the error
messages since we actually don't use these methods (yet).

The next commit will contains the scm_errors_test.go that checks
that the appropriate error message is returned.