refactor(scm): make userErr be an error proper
The previous use of Msg to represent the user error message to be
exposed to the client now uses a proper error type instead of the
weird Msg construction. This is API compatible with the M() helper
function to create &userError types.

We could consider exporting this as a UserError type and let it
be part of the error chain and extract it from the SCMError using
the errors.As(err, &userErr) approach to directly get it for use
in the parseSCMError function.