refactor(scm): SCMError now only keeps a single error chain
We no longer separate between error and userError; instead we keep
all errors in a single chain; this makes it easier to work with.
However, we still need to search with errors.As(err, &scmError)
in the parseSCMError() func; not sure why it isn't able to find
the UserError directly and still use the root SCMError inside the
if statement.