fix(scm): UserError now can be extracted from error chain
The previous commit introduced a bug that prevented the error
chain from keeping error information about the error type.

This adds a test to detect this problem should it reappear, and
fixes it by using %w for both the old and the new error when
chaining errors.

This also fixes the limitation that regular errors would overwrite
existing errors; it now adds regular errors into the chain as well;
not just the SCMError, UserError and strings.