fix(scm): update error messages and tests and support debugErrMsg
This adjusts the error messages and corresponding tests to work
with the new chained variant (where both user errors and other
errors are in the same chain).

This also simplifies the tests using a common test helper:
chkErrMsg that will log appropriate output when a test fails,
including a full line view of the error and user error (for
convenience when updating the tests.)

The chkErrMsg helper also can emit the error messages by providing
the DEBUG_ERR_MSG env variable. This is handy for seeing all the
error messages currently produced by the test.