test(web): replace CreateCourse RPC calls with qtest.CreateCourse
In preparation for removal of the CreateCourse RPC, this removes
its use in most tests; turns out most cases do not depend on
interacting with the SCM. The exception is TestEnrollmentProcess
which uses UpdateEnrollments that depends on interacting with the
SCM. However, it was easy to fix by using scm.WithMockCourses()
to initialize the SCM so that the enrollment process would still
be testable.

TestNewCourseExistingRepos has not been updated yet; it will be
removed and replaced by another test in the hooks package in the
next commit, since we no longer have a way to check the returned
error from scm.CreateCourse in the CreateCourse RPC; we no longer
have a way to pass it on to the frontend. Instead we log this fact
during the github app install event handler for the quickfeed app.
(hooks/installation.go:handleInstallationCreated).