test(web|hooks): add TestAlreadyExistingCourse
This replaces web.TestNewCourseExistingRepos with the new test
in hooks. This is since we no longer have a way to check that
the returned error from scm.CreateCourse is scm.ErrAlreadyExists
error. Since we no longer have a frontend to display this error
message to (since we no longer use the web.CreateCourse RPC),
we instead log this fact during the github app install event
handling. This commit is only adding this test and removing the
old ErrAlreadyExists test (TestNewCourseExistingRepos).