test(interceptor): remove CreateCourse token update func
This removes the token update function for the CreateCourse RPC
method; this was used to upgrade the signed in user to get the
teacher role in the new course.

This also removes a check in the TestRefreshTokens related to the
CreateCourse RPC; this check is no longer relevant since the
admin is not being promoted to teacher via the CreateCourse RPC
anymore; this is instead handled by the handleInstallationCreated()
in hooks/installation.go.