chore: use qtest.CreateCourse instead of err = db.CreateCourse
In tests we use a fair amount of both qtest.CreateCourse() and
err = db.CreateCourse() along with the err check and t.Fatal(err).
We can simplify most tests by using qtest.CreateCourse() all over
instead of the current mixture of both. Note that there are some
cases where we need to check the err return value, which requires
that we use the traditional db.CreateCourse().