fix(scm): add unwrap iface to force {User|SCM}Error impl Unwrap()
This adds an unwrap interface to force the User and SCM errors to
implement these methods. While it is a bit late now, since they
already implement Unwrap(), it can help others (and me) looking at
this example and remember that to use the errors.As function to
check for a specific error type, the error type must implement the
Unwrap() method.