refactor(qf): replace IsApproved with SubmissionStatus
This replaces IsApproved method with SubmissionStatus which returns
a slice of grades as before; the signatures are compatible.
However, the SubmissionStatus ignores the validation checks since
we do similar checks in the CreateSubmission method in the database
layer; although we don't actually check that the submission matches
the assignment's group vs user assignment. If we want to detect
that we may need to add logic elsewhere. The existing logic only
marks the grade submission status as none (to ignore the submission)
for a group submission on a user assignment and vice versa.