refactor: simplified `fetchUserData` such that it does not preemptively fetch *all* data
Course specific data is now fetched when a user navigates to a specific course.

User groups are included in `getEnrollmentsByUser` and do not need to be fetched twice. getGroup is removed as a consequence.

Course groups are also included in `GetCourse` if the user is a teacher, and do not need to be fetched separately.