feat(kit/score): NewSocketRegistry and NewSocket() for score comms
This adds NewSocket() for server-side (quickfeed server) to create
a new unix domain socket for sending score JSON objects on a
separate channel from the vanilla stdout we currently use.
To take advantage of this, the main_test.go file need to use the
NewSocketRegistry() and instead of doing defer sc.Print(t), we must
do defer scores.PrintToSocket(t, sc).

Note, we can avoid doing NewSocketRegistry by adding its code into
the current NewRegistry() and ignore the error and fall back to the
stdout version if the domain socket isn't available. I'm undecided
if we want to do that or not. Letting it brew for a while.