refactor(kit/score): NewRegistry create socket; add Print(t, sc)
This migrates the NewSocketRegistry() version in registry_socket.go
to NewRegistry(), allowing to fall back to stdout score reporting
if we cannot connect to the unix domain socket. Note this currently
does report this error. It is not clear if we should or not.

Right now, this version is compatible with the existing tests.
However, existing tests will need to use scores.Print(t, sc) instead
of sc.Print(t) to take advantage of the socket based score reporting.