fix(main): add -secret flag to quickfeed server
This will generate a random string for the QUICKFEED_AUTH_SECRET,
which is the secret used to sign JWT tokens. The generated secret
will be saved in the .env file for future reuse. To rotate the
secret, run the server with the -secret flag again, or pass a
custom secret via the environment variable QUICKFEED_AUTH_SECRET
from the shell. Rotating the secret will log out users from the
server.