feat: replace score.Score with ExpectedTests in Assignment
This walks back the first idea to reuse the zero-th submission
entry in the database; it complicated things too much and I
think it will be harder to maintain such duplication of use
of the existing Assignment.Submissions field.

Instead, this creates a new TestInfo message that encapsulate details
about each test, including fields for ID, AssignmentID (foreign key),
TestName, MaxScore, Weight, and Details. Only the TestName, MaxScore,
and Weight are currently populated in the existing JSON files, but
we could possibly add Details if we find that useful somehow?!

A slice of these such TestInfo entries are filled into ExpectedTests.

We still need to fix the database handling of these ExpectedTests.