fix(types.proto): add unique index for AssignmentID and TestName
This removes the previous foreign key reference, which gorm can
determine automatically. Moreover, it adds a unique index for the
AssignmentID and TestName.

The purpose of the index is to speed up queries that filter or sort
by the indexed fields, such as looking up tests by AssignmentID and
TestName. Declaring the index as unique ensures that no two TestInfo
entries can have the same combination of AssignmentID and TestName,
effectively enforcing a constraint that each test name must be unique
within a given assignment. This helps maintain data integrity and
improves query performance.