fix: add -coverpkg in test command
This ensures that codecov will see all code that is covered, even if we
do not have specific tests for the relevant methods. For example, we
don't need to duplicate the same tests in the database layer when we
already have tests in the web layer for the same database methods.