feat: add go.mod and go.sum for build context in testdata
The implementation of readTestsRepositoryContent assumes that
go.mod and go.sum files are added in the tests repo somewhere;
previously we did not add go.* files to the tests repo, relying
on the assignments repo to provide this. However, we now require
that they be provided for build context during the docker build
process, since otherwise it would be a bit difficult to support
this context-based building based only on the tests repository.

We will need to test this extensively to avoid causing problems
for the compiler during tests; that is, the compiler may become
distracted by the go.* files in the scripts folder.