test: add TestIsValid to check validation logic
This checks all implementations of the validator interface (IsValid).
Also updated doc comments for the various IsValid methods to be
more consistent and made various other adjustments to the logic
for clarity (readability of the code). Moreover, I also added a
check to the Enrollment.IsValid() implementation to limit the
acceptable Status range to (NONE, TEACHER).